Рефакторинг и доп компонент
