+++ project/src/components/room/nearby-blok/nearby-rooms.tsx
@@ -1,28 +1,14 @@
 import {memo} from 'react';
 
-import PlaceCard from '../place-card/place-card';
-import Spinner from '../spinner/spinner';
+import PlaceCard from '../../place-card/place-card';
 
-import {useAppSelector} from '../../hooks';
+import {useAppSelector} from '../../../hooks';
 
 
 function NearbyRooms(): JSX.Element {
   // eslint-disable-next-line no-console
   console.log('nearby');
   const serverNearbyOffers = useAppSelector((state) => state.serverNearbyOffers);
-  const isNearbyOffersDataLoading = useAppSelector((state) => state.isNearbyOffersDataLoading);
-
-  if (isNearbyOffersDataLoading) {
-    const spinnerSize = {
-      width: 1000,
-      height: 424,
-    };
-    return (
-      <section className="near-places places">
-        <Spinner spinnerSize={[spinnerSize.height, spinnerSize.width]}/>
-      </section>
-    );
-  }
 
   return (
     <section className="near-places places">
