+++ project/src/components/room/room-review/add-review-form.tsx
@@ -4,11 +4,11 @@ import {useEffect} from 'react';
 //import {memo} from 'react';
 import { useParams } from 'react-router';
 
-import {NewComment} from '../../types/comment';
+import {NewComment} from '../../../types/comment';
 
-import {fetchPostCommentAction} from '../../store/api-actions';
-import {useAppDispatch, useAppSelector} from '../../hooks';
-import {setCommentSubmutAction} from '../../store/action';
+import {fetchPostCommentAction} from '../../../store/api-actions';
+import {useAppDispatch, useAppSelector} from '../../../hooks';
+import {setCommentSubmutAction} from '../../../store/action';
 
 const COMMENT_LENGTH = {
   min: 50,
