+++ project/src/components/room/room-review/review.tsx
@@ -1,4 +1,4 @@
-import {Comment} from '../../types/comment';
+import {Comment} from '../../../types/comment';
 
 type ReviewProps = {
   comment: Comment;
